[11/17/2022] [00:31] Update System has been finnished. New Scripts and bug with same max and min fixed
So finally now you can send to the number 'Update' to recieve the current forecast.

New Scripts for AddUser.py to add a user to the db, and MigrateFromUserToUsernames.py to get all the usernames to be in usernames table.

Finally Bug Fix Max/Min Temp now show different data